fix: make index a map, add comparator for cells
